test(token_usage): restructure and consolidate unit tests into router test module
